library migrated to supabase is now fetched
