--- salah-app/src/app/library/library.contant.ts
@@ -1,996 +1,0 @@
-import { IslamicLibrary } from "../model/islamic-library.model";
-
-export const library: IslamicLibrary[] = [
-    {
-        name: "Qur'an Full",
-        pdfName: 'quran_13_liner_color_coded.pdf',
-        category: 'quran',
-        storageKey: 'quranPage'
-    },
-    {
-        name: "Qur'an Translation English",
-        pdfName: 'quran_translation_english.pdf',
-        category: 'quran',
-        storageKey: 'quranEnglishPage'
-    },
-    {
-        name: "Para 01 - Alif Lam Meem (الٓمٓ)",
-        pdfName: 'para01.pdf',
-        category: 'quran',
-        storageKey: 'para01Page'
-    },
-    {
-        name: "Para 02 - Sayaqool (سَيَقُولُ)",
-        pdfName: 'para02.pdf',
-        category: 'quran',
-        storageKey: 'para02Page'
-    },
-    {
-        name: "Para 03 - Tilka ar-Rusul (تِلْكَ ٱلرُّسُلُ)",
-        pdfName: 'para03.pdf',
-        category: 'quran',
-        storageKey: 'para03Page'
-    },
-    {
-        name: "Para 04 - Lan Tana Loo (لَن تَنَالُوا۟)",
-        pdfName: 'para04.pdf',
-        category: 'quran',
-        storageKey: 'para04Page'
-    },
-    {
-        name: "Para 05 - Wal Mohsanat (وَٱلْمُحْصَنَٰتُ)",
-        pdfName: 'para05.pdf',
-        category: 'quran',
-        storageKey: 'para05Page'
-    },
-    {
-        name: "Para 06 - La Yuhibbullah (لَا يُحِبُّ ٱللَّهُ)",
-        pdfName: 'para06.pdf',
-        category: 'quran',
-        storageKey: 'para06Page'
-    },
-    {
-        name: "Para 07 - Wa Iza Sami'u (وَإِذَا سَمِعُوا۟)",
-        pdfName: 'para07.pdf',
-        category: 'quran',
-        storageKey: 'para07Page'
-    },
-    {
-        name: "Para 08 - Wa Lau Annana (وَلَوْ أَنَّنَا)",
-        pdfName: 'para08.pdf',
-        category: 'quran',
-        storageKey: 'para08Page'
-    },
-    {
-        name: "Para 09 - Qad Aflaha (قَدْ أَفْلَحَ)",
-        pdfName: 'para09.pdf',
-        category: 'quran',
-        storageKey: 'para09Page'
-    },
-    {
-        name: "Para 10 - Wa A'lamu (وَٱعْلَمُوٓا۟)",
-        pdfName: 'para10.pdf',
-        category: 'quran',
-        storageKey: 'para10Page'
-    },
-    {
-        name: "Para 11 - Ya'tazerun (يَعْتَذِرُونَ)",
-        pdfName: 'para11.pdf',
-        category: 'quran',
-        storageKey: 'para11Page'
-    },
-    {
-        name: "Para 12 - Wa Mamin Da'abat (وَمَا مِن دَآبَّةٍۢ)",
-        pdfName: 'para12.pdf',
-        category: 'quran',
-        storageKey: 'para12Page'
-    },
-    {
-        name: "Para 13 - Wa Ma Ubrioo (وَمَآ أُبَرِّئُ)",
-        pdfName: 'para13.pdf',
-        category: 'quran',
-        storageKey: 'para13Page'
-    },
-    {
-        name: "Para 14 - Rubama (رُّبَمَا)",
-        pdfName: 'para14.pdf',
-        category: 'quran',
-        storageKey: 'para14Page'
-    },
-    {
-        name: "Para 15 - Subhanalladhi (سُبْحَٰنَ ٱلَّذِى)",
-        pdfName: 'para15.pdf',
-        category: 'quran',
-        storageKey: 'para15Page'
-    },
-    {
-        name: "Para 16 - Qal Alam (قَالَ أَلَمْ)",
-        pdfName: 'para16.pdf',
-        category: 'quran',
-        storageKey: 'para16Page'
-    },
-    {
-        name: "Para 17 - Iqtarabat (ٱقْتَرَبَتِ)",
-        pdfName: 'para17.pdf',
-        category: 'quran',
-        storageKey: 'para17Page'
-    },
-    {
-        name: "Para 18 - Qadd Aflaha (قَدْ أَفْلَحَ)",
-        pdfName: 'para18.pdf',
-        category: 'quran',
-        storageKey: 'para18Page'
-    },
-    {
-        name: "Para 19 - Wa Qalalladhina (وَقَالَ ٱلَّذِينَ)",
-        pdfName: 'para19.pdf',
-        category: 'quran',
-        storageKey: 'para19Page'
-    },
-    {
-        name: "Para 20 - Amman Khalaq (أَمَّنْ خَلَقَ)",
-        pdfName: 'para20.pdf',
-        category: 'quran',
-        storageKey: 'para20Page'
-    },
-    {
-        name: "Para 21 - Utlu Ma Oohi (ٱتْلُ مَآ أُوحِىَ)",
-        pdfName: 'para21.pdf',
-        category: 'quran',
-        storageKey: 'para21Page'
-    },
-    {
-        name: "Para 22 - Wa Manyaqnut (وَمَن يَقْنُتْ)",
-        pdfName: 'para22.pdf',
-        category: 'quran',
-        storageKey: 'para22Page'
-    },
-    {
-        name: "Para 23 - Wa Mali (وَمَا لِىَ)",
-        pdfName: 'para23.pdf',
-        category: 'quran',
-        storageKey: 'para23Page'
-    },
-    {
-        name: "Para 24 - Faman Azlam (فَمَنْ أَظْلَمُ)",
-        pdfName: 'para24.pdf',
-        category: 'quran',
-        storageKey: 'para24Page'
-    },
-    {
-        name: "Para 25 - Elahe Yuruddo (إِلَيْهِ يُرَدُّ)",
-        pdfName: 'para25.pdf',
-        category: 'quran',
-        storageKey: 'para25Page'
-    },
-    {
-        name: "Para 26 - Ha Meem (حم)",
-        pdfName: 'para26.pdf',
-        category: 'quran',
-        storageKey: 'para26Page'
-    },
-    {
-        name: "Para 27 - Qala Fama Khatbukum (قَالَ فَمَا خَطْبُكُمْ)",
-        pdfName: 'para27.pdf',
-        category: 'quran',
-        storageKey: 'para27Page'
-    },
-    {
-        name: "Para 28 - Qadd Sami Allah (قَدْ سَمِعَ اللهُ)",
-        pdfName: 'para28.pdf',
-        category: 'quran',
-        storageKey: 'para28Page'
-    },
-    {
-        name: "Para 29 - Tabarakalladhi (تَبَٰرَكَ ٱلَّذِى)",
-        pdfName: 'para29.pdf',
-        category: 'quran',
-        storageKey: 'para29Page'
-    },
-    {
-        name: "Para 30 - Amma Yatasa'aloon (عَمَّ يَتَسَآءَلُونَ)",
-        pdfName: 'para30.pdf',
-        category: 'quran',
-        storageKey: 'para30Page'
-    },
-    {
-        name: "the sealed nectar",
-        pdfName: 'the_sealed_nectar.pdf',
-        category: 'prophet_muhammad_(ﷺ)',
-        storageKey: 'the_sealed_nectar'
-    },
-    {
-        name: "the life of the prophet muhammad - volume 1",
-        pdfName: 'the_life_of_the_prophet_muhammad_volume_1.pdf',
-        category: 'prophet_muhammad_(ﷺ)',
-        storageKey: 'lifeOfProphet1'
-    },
-    {
-        name: "the life of the prophet muhammad - volume 2",
-        pdfName: 'the_life_of_the_prophet_muhammad_volume_2.pdf',
-        category: 'prophet_muhammad_(ﷺ)',
-        storageKey: 'lifeOfProphet2'
-    },
-    {
-        name: "the life of the prophet muhammad - volume 3",
-        pdfName: 'the_life_of_the_prophet_muhammad_volume_3.pdf',
-        category: 'prophet_muhammad_(ﷺ)',
-        storageKey: 'lifeOfProphet3'
-    },
-    {
-        name: "the life of the prophet muhammad - volume 4",
-        pdfName: 'the_life_of_the_prophet_muhammad_volume_4.pdf',
-        category: 'prophet_muhammad_(ﷺ)',
-        storageKey: 'lifeOfProphet4'
-    },
-    {
-        name: "surah 1 - al fatihah",
-        pdfName: 'surah-1-al-fatihah.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir1'
-    },
-    {
-        name: "surah 2 - al baqara part i",
-        pdfName: 'surah-2-al-baqara-part-i.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir2_1'
-    },
-    {
-        name: "surah 2 - al baqara part ii",
-        pdfName: 'surah-2-al-baqara-part-ii.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir2_2'
-    },
-    {
-        name: "surah 3 - aali imran",
-        pdfName: 'surah-3-aali-imran.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir3'
-    },
-    {
-        name: "surah 4 - an nisa",
-        pdfName: 'surah-4-an-nisa.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir4'
-    },
-    {
-        name: "surah 5 - al maidah",
-        pdfName: 'surah-5-al-maidah.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir5'
-    },
-    {
-        name: "surah 6 - al anaam",
-        pdfName: 'surah-6-al-anaam.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir6'
-    },
-    {
-        name: "surah 7 - al aaraf",
-        pdfName: 'surah-7-al-aaraf.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir7'
-    },
-    {
-        name: "surah 8 - al anfal",
-        pdfName: 'surah-8-al-anfal.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir8'
-    },
-    {
-        name: "surah 9 - at tawbah",
-        pdfName: 'surah-9-at-tawbah.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir9'
-    },
-    {
-        name: "surah 10 - yunus",
-        pdfName: 'surah-10-yunus.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir10'
-    },
-
-    {
-        name: "surah 11 - hud",
-        pdfName: 'surah-11-hud.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir11'
-    },
-
-    {
-        name: "surah 12 - yusuf",
-        pdfName: 'surah-12-yusuf.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir12'
-    },
-    {
-        name: "surah 13 - ar raad",
-        pdfName: 'surah-13-ar-raad.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir13'
-    },
-    {
-        name: "surah 14 - ibrahim",
-        pdfName: 'surah-14-ibrahim.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir14'
-    },
-    {
-        name: "surah 15 - al hijr",
-        pdfName: 'surah-15-al-hijr.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir15'
-    },
-    {
-        name: "surah 16 - an nahl",
-        pdfName: 'surah-16-an-nahl.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir16'
-    },
-    {
-        name: "surah 17 - al isra",
-        pdfName: 'surah-17-al-isra.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir17'
-    },
-    {
-        name: "surah 18 - al kahf",
-        pdfName: 'surah-18-al-kahf.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir18'
-    },
-    {
-        name: "surah 19 - maryam",
-        pdfName: 'surah-19-maryam.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir19'
-    },
-    {
-        name: "surah 20 - taha",
-        pdfName: 'surah-20-taha.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir20'
-    },
-    {
-        name: "surah 21 - al anbya",
-        pdfName: 'surah-21-al-anbya.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir21'
-    },
-    {
-        name: "surah 22 - al haj",
-        pdfName: 'surah-22-al-haj.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir22'
-    },
-    {
-        name: "surah 23 - al muminun",
-        pdfName: 'surah-23-al-muminun.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir23'
-    },
-    {
-        name: "surah 24 - an nur",
-        pdfName: 'surah-24-an-nur.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir24'
-    },
-    {
-        name: "surah 25 - al furqan",
-        pdfName: 'surah-25-al-furqan.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir25'
-    },
-    {
-        name: "surah 26 - ash shuara",
-        pdfName: 'surah-26-ash-shuara.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir26'
-    },
-    {
-        name: "surah 27 - an naml",
-        pdfName: 'surah-27-an-naml.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir27'
-    },
-    {
-        name: "surah 28 - al qasas",
-        pdfName: 'surah-28-al-qasas.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir28'
-    },
-    {
-        name: "surah 29 - al ankabut",
-        pdfName: 'surah-29-al-ankabut.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir29'
-    },
-
-    {
-        name: "surah 30 - ar rum",
-        pdfName: 'surah-30-ar-rum.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir30'
-    },
-    {
-        name: "surah 31 - luqman",
-        pdfName: 'surah-31-luqman.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir31'
-    },
-    {
-        name: "surah 32 - as sajdah",
-        pdfName: 'surah-32-as-sajdah.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir32'
-    },
-    {
-        name: "surah 33 - al ahzab",
-        pdfName: 'surah-33-al-ahzab.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir33'
-    },
-    {
-        name: "surah 34 - saba",
-        pdfName: 'surah-34-saba.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir34'
-    },
-    {
-        name: "surah 35 - fatir",
-        pdfName: 'surah-35-fatir.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir35'
-    },
-    {
-        name: "surah 36 - ya sin",
-        pdfName: 'surah-36-ya-sin.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir36'
-    },
-    {
-        name: "surah 37 - as saffat",
-        pdfName: 'surah-37-as-saffat.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir37'
-    },
-    {
-        name: "surah 38 - sad",
-        pdfName: 'surah-38-sad.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir38'
-    },
-    {
-        name: "surah 39 - az zumar",
-        pdfName: 'surah-39-az-zumar.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir39'
-    },
-
-    {
-        name: "surah 40 - ghafir",
-        pdfName: 'surah-40-ghafir.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir40'
-    },
-    {
-        name: "surah 41 - fussilat",
-        pdfName: 'surah-41-fussilat.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir41'
-    },
-    {
-        name: "surah 42 - ash shuraa",
-        pdfName: 'surah-42-ash-shuraa.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir42'
-    },
-    {
-        name: "surah 43 - az zukhruf",
-        pdfName: 'surah-43-az-zukhruf.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir43'
-    }, {
-        name: "surah 44 - ad dukhan",
-        pdfName: 'surah-44-ad-dukhan.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir44'
-    },
-    {
-        name: "surah 45 - al jathiyah",
-        pdfName: 'surah-45-al-jathiyah.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir45'
-    },
-    {
-        name: "surah 46 - al ahqaf",
-        pdfName: 'surah-46-al-ahqaf.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir46'
-    },
-    {
-        name: "surah 47 - muhammad",
-        pdfName: 'surah-47-muhammad.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir47'
-    },
-    {
-        name: "surah 48 - al fath",
-        pdfName: 'surah-48-al-fath.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir48'
-    },
-    {
-        name: "surah 49 - al hujurat",
-        pdfName: 'surah-49-al-hujurat.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir49'
-    },
-    {
-        name: "surah 50 - qaf",
-        pdfName: 'surah-50-qaf.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir50'
-    },
-    {
-        name: "surah 51 - adh dhariyat",
-        pdfName: 'surah-51-adh-dhariyat.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir51'
-    },
-    {
-        name: "surah 52 - at tur",
-        pdfName: 'surah-52-at-tur.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir52'
-    },
-    {
-        name: "surah 53 - an najm",
-        pdfName: 'surah-53-an-najm.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir53'
-    },
-    {
-        name: "surah 54 - al qamar",
-        pdfName: 'surah-54-al-qamar.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir54'
-    },
-    {
-        name: "surah 55 - ar rahman",
-        pdfName: 'surah-55-ar-rahman.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir55'
-    },
-    {
-        name: "surah 56 - al waqiah",
-        pdfName: 'surah-56-al-waqiah.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir56'
-    },
-    {
-        name: "surah 57 - al hadid",
-        pdfName: 'surah-57-al-hadid.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir57'
-    },
-    {
-        name: "surah 58 - al mujadila",
-        pdfName: 'surah-58-al-mujadila.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir58'
-    },
-    {
-        name: "surah 59 - al hashr",
-        pdfName: 'surah-59-al-hashr.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir59'
-    },
-    {
-        name: "surah 60 - al mumtahanah",
-        pdfName: 'surah-60-al-mumtahanah.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir60'
-    },
-    {
-        name: "surah 61 - as saf",
-        pdfName: 'surah-61-as-saf.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir61'
-    },
-    {
-        name: "surah 62 - al jumuah",
-        pdfName: 'surah-62-al-jumuah.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir62'
-    },
-    {
-        name: "surah 63 - al munafiqun",
-        pdfName: 'surah-63-al-munafiqun.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir63'
-    },
-    {
-        name: "surah 64 - at taghabun",
-        pdfName: 'surah-64-at-taghabun.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir64'
-    },
-    {
-        name: "surah 65 - at talaq",
-        pdfName: 'surah-65-at-talaq.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir65'
-    },
-    {
-        name: "surah 66 - at tahrim",
-        pdfName: 'surah-66-at-tahrim.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir66'
-    },
-    {
-        name: "surah 67 - al mulk",
-        pdfName: 'surah-67-al-mulk.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir67'
-    },
-    {
-        name: "surah 68 - al qalam",
-        pdfName: 'surah-68-al-qalam.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir68'
-    },
-    {
-        name: "surah 69 - al haqqah",
-        pdfName: 'surah-69-al-haqqah.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir69'
-    },
-    {
-        name: "surah 70 - al maarij",
-        pdfName: 'surah-70-al-maarij.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir70'
-    },
-    {
-        name: "surah 71 - nuh",
-        pdfName: 'surah-71-nuh.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir71'
-    },
-    {
-        name: "surah 72 - al jinn",
-        pdfName: 'surah-72-al-jinn.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir72'
-    },
-    {
-        name: "surah 73 - al muzzammil",
-        pdfName: 'surah-73-al-muzzammil.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir73'
-    },
-    {
-        name: "surah 74 - al muddaththir",
-        pdfName: 'surah-74-al-muddaththir.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir74'
-    },
-    {
-        name: "surah 75 - al qiyamah",
-        pdfName: 'surah-75-al-qiyamah.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir75'
-    },
-    {
-        name: "surah 76 - al insan",
-        pdfName: 'surah-76-al-insan.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir76'
-    },
-    {
-        name: "surah 77 - al mursalat",
-        pdfName: 'surah-77-al-mursalat.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir77'
-    },
-    {
-        name: "surah 78 - an naba",
-        pdfName: 'surah-78-an-naba.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir78'
-    },
-    {
-        name: "surah 79 - an naziaat",
-        pdfName: 'surah-79-an-naziaat.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir79'
-    },
-    {
-        name: "surah 80 - abasa",
-        pdfName: 'surah-80-abasa.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir80'
-    },
-    {
-        name: "surah 81 - at takwir",
-        pdfName: 'surah-81-at-takwir.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir81'
-    },
-    {
-        name: "surah 82 - al infitar",
-        pdfName: 'surah-82-al-infitar.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir82'
-    },
-    {
-        name: "surah 83 - al mutaffifin",
-        pdfName: 'surah-83-al-mutaffifin.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir83'
-    },
-    {
-        name: "surah 84 - al inshiqaq",
-        pdfName: 'surah-84-al-inshiqaq.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir84'
-    },
-    {
-        name: "surah 85 - al buruj",
-        pdfName: 'surah-85-al-buruj.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir85'
-    },
-    {
-        name: "surah 86 - at tariq",
-        pdfName: 'surah-86-at-tariq.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir86'
-    },
-    {
-        name: "surah 87 - al aala",
-        pdfName: 'surah-87-al-aala.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir87'
-    },
-    {
-        name: "surah 88 - al ghashiyah",
-        pdfName: 'surah-88-al-ghashiyah.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir88'
-    },
-    {
-        name: "surah 89 - al fajr",
-        pdfName: 'surah-89-al-fajr.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir89'
-    },
-    {
-        name: "surah 90 - al balad",
-        pdfName: 'surah-90-al-balad.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir90'
-    },
-    {
-        name: "surah 91 - ash shams",
-        pdfName: 'surah-91-ash-shams.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir91'
-    },
-    {
-        name: "surah 92 - al layl",
-        pdfName: 'surah-92-al-layl.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir92'
-    },
-    {
-        name: "surah 93 - ad duhaa",
-        pdfName: 'surah-93-ad-duhaa.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir93'
-    },
-    {
-        name: "surah 94 - ash sharh",
-        pdfName: 'surah-94-ash-sharh.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir94'
-    },
-    {
-        name: "surah 95 - at tin",
-        pdfName: 'surah-95-at-tin.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir95'
-    },
-    {
-        name: "surah 96 - al alaq",
-        pdfName: 'surah-96-al-alaq.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir96'
-    },
-    {
-        name: "surah 97 - al qadr",
-        pdfName: 'surah-97-al-qadr.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir97'
-    }, {
-        name: "surah 98 - al bayyinah",
-        pdfName: 'surah-98-al-bayyinah.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir98'
-    },
-    {
-        name: "surah 99 - az zalzalah",
-        pdfName: 'surah-99-az-zalzalah.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir99'
-    },
-    {
-        name: "surah 100 - al aadiyat",
-        pdfName: 'surah-100-al-aadiyat.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir100'
-    },
-    {
-        name: "surah 101 - al qariah",
-        pdfName: 'surah-101-al-qariah.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir101'
-    },
-    {
-        name: "surah 102 - at takathur",
-        pdfName: 'surah-102-at-takathur.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir102'
-    },
-    {
-        name: "surah 103 - al asr",
-        pdfName: 'surah-103-al-asr.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir103'
-    },
-    {
-        name: "surah 104 - al humazah",
-        pdfName: 'surah-104-al-humazah.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir104'
-    },
-    {
-        name: "surah 105 - al fil",
-        pdfName: 'surah-105-al-fil.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir105'
-    },
-    {
-        name: "surah 106 - quraysh",
-        pdfName: 'surah-106-quraysh.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir106'
-    },
-    {
-        name: "surah 107 - al maun",
-        pdfName: 'surah-107-al-maun.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir107'
-    },
-    {
-        name: "surah 108 - al kawthar",
-        pdfName: 'surah-108-al-kawthar.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir108'
-    },
-    {
-        name: "surah 109 - al kafirun",
-        pdfName: 'surah-109-al-kafirun.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir109'
-    },
-    {
-        name: "surah 110 - an nasr",
-        pdfName: 'surah-110-an-nasr.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir110'
-    },
-    {
-        name: "surah 111 - al masad",
-        pdfName: 'surah-111-al-masad.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir111'
-    },
-    {
-        name: "surah 112 - al ikhlas",
-        pdfName: 'surah-112-al-ikhlas.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir112'
-    },
-    {
-        name: "surah 113 - al falaq",
-        pdfName: 'surah-113-al-falaq.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir113'
-    },
-    {
-        name: "surah 114 - an nas",
-        pdfName: 'surah-114-an-nas.pdf',
-        category: 'tafsir_al_quran_ibn_kathir',
-        storageKey: 'kathir114'
-    },
-    {
-        name: "Ishraq Index",
-        pdfName: 'ishraq_index.pdf',
-        category: 'tafsir_al_Quran_ishraq',
-        storageKey: 'ishraqIndex'
-    },
-    {
-        name: "ishraq preface complete",
-        pdfName: 'ishraq_preface_complete.pdf',
-        category: 'tafsir_al_Quran_ishraq',
-        storageKey: 'ishraqPreface'
-    },
-    {
-        name: "ishraq volume 1",
-        pdfName: 'ishraq_volume_1.pdf',
-        category: 'tafsir_al_Quran_ishraq',
-        storageKey: 'ishraq1'
-    },
-    {
-        name: "ishraq volume 2",
-        pdfName: 'ishraq_volume_2.pdf',
-        category: 'tafsir_al_Quran_ishraq',
-        storageKey: 'ishraq2'
-    },
-    {
-        name: "ishraq volume 3",
-        pdfName: 'ishraq_volume_3.pdf',
-        category: 'tafsir_al_Quran_ishraq',
-        storageKey: 'ishraq3'
-    },
-    {
-        name: "ishraq volume 4",
-        pdfName: 'ishraq_volume_1.pdf',
-        category: 'tafsir_al_Quran_ishraq',
-        storageKey: 'ishraq4'
-    },
-    {
-        name: "ishraq volume 5",
-        pdfName: 'ishraq_volume_5.pdf',
-        category: 'tafsir_al_Quran_ishraq',
-        storageKey: 'ishraq5'
-    },
-    {
-        name: "ishraq volume 6",
-        pdfName: 'ishraq_volume_6.pdf',
-        category: 'tafsir_al_Quran_ishraq',
-        storageKey: 'ishraq6'
-    },
-    {
-        name: "ishraq volume 7",
-        pdfName: 'ishraq_volume_7.pdf',
-        category: 'tafsir_al_Quran_ishraq',
-        storageKey: 'ishraq7'
-    },
-    {
-        name: "ishraq volume 8",
-        pdfName: 'ishraq_volume_1.pdf',
-        category: 'tafsir_al_Quran_ishraq',
-        storageKey: 'ishraq8'
-    },
-    {
-        name: "after prophet",
-        pdfName: 'After_the_Prophet-The_Epic_Story_of_the_Shia-Sunni_Split_in_Islam.pdf',
-        category: 'islam',
-        storageKey: 'afterProphet'
-    },
-    {
-        name: "stories of the prophets",
-        pdfName: 'stories_of_the_prophets.pdf',
-        category: 'islam',
-        storageKey: 'storiesOfProphets'
-    },
-    {
-        name: "Riyad-Us-Saliheen",
-        pdfName: 'Riyad_Us-Saliheen.pdf',
-        category: 'islam',
-        storageKey: 'riyadussaliheen'
-    }
-];
